+++ Changelog.rst
@@ -3,6 +3,14 @@ Version NEXTVERSION
 
 **2026-??-??**
 
+* New methods to convert to `xarray`: `cf.Field.to_xarray` and
+  `cf.Domain.to_xarray`
+  (https://github.com/NCAS-CMS/cfdm/issues/394)
+* New output format for `cfdm.write` that creates an `xarray` dataset
+  in memory: ``'XARRAY'``
+  (https://github.com/NCAS-CMS/cfdm/issues/394)
+* Write UGRID datasets with `cfdm.write`
+  (https://github.com/NCAS-CMS/cfdm/issues/271)
 * New keyword to `cfdm.read`: ``filesystem``
   (https://github.com/NCAS-CMS/cfdm/issues/397)
 * New keyword parameter to `cfdm.Data.compute`: ``persist``
@@ -31,6 +39,7 @@ Version NEXTVERSION
   different ``formula_terms``
   (https://github.com/NCAS-CMS/cfdm/issues/380).
 * New dependency: ``pyfive>=1.1.1``
+* New optional dependency: ``xarray>=2026.2.0``
 * Changed dependency: ``h5netcdf>=1.8.0``
 
 ----